Update patcher to use modern browser controls Fix bug with fishing rods Code cleanup
