+++ Server/MirObjects/MapObject.cs
@@ -60,10 +60,6 @@ namespace Server.MirObjects
         public byte Light;
         public int AttackSpeed;
 
-        public int CurrentHandWeight,
-                   CurrentWearWeight,
-                   CurrentBagWeight;
-
         public long CellTime, BrownTime, PKPointTime, LastHitTime, EXPOwnerTime;
         public Color NameColour = Color.White;
         
@@ -121,8 +117,6 @@ namespace Server.MirObjects
                 _sneakingActive = value;
 
                 Observer = _sneakingActive;
-
-                //CurrentMap.Broadcast(new S.ObjectSneaking { ObjectID = ObjectID, SneakingActive = value }, CurrentLocation);
             }
         }
 
@@ -158,7 +152,6 @@ namespace Server.MirObjects
         public List<PlayerObject> GroupMembers;
 
         public virtual AttackMode AMode { get; set; }
-
         public virtual PetMode PMode { get; set; }
 
         private bool _inSafeZone;
@@ -235,6 +228,7 @@ namespace Server.MirObjects
         {
             for (int i = 0; i < Buffs.Count; i++)
             {
+                if (Buffs[i].ObjectID == 0) continue;
                 if (!Buffs[i].Properties.HasFlag(BuffProperty.PauseInSafeZone)) continue;
 
                 if (InSafeZone)
@@ -297,8 +291,8 @@ namespace Server.MirObjects
         {
             if (Race == ObjectType.Merchant)
             {
-                NPCObject NPC = (NPCObject)this;
-                NPC.CheckVisible(player, true);
+                NPCObject npc = (NPCObject)this;
+                npc.CheckVisible(player, true);
                 return;
             }
 
@@ -897,6 +891,7 @@ namespace Server.MirObjects
         }
 
         public abstract void Die();
+
         public abstract int Pushed(MapObject pusher, MirDirection dir, int distance);
 
         public bool IsMember(MapObject member)
@@ -918,7 +913,7 @@ namespace Server.MirObjects
             {
                 if (this is PlayerObject)
                 {
-                    PlayerObject player = (PlayerObject)this;
+                    var player = (PlayerObject)this;
                     player.Enqueue(new S.InTrapRock { Trapped = value });
                 }
             }
@@ -983,189 +978,6 @@ namespace Server.MirObjects
             Time = reader.ReadInt64();
             TickTime = reader.ReadInt64();
             TickSpeed = reader.ReadInt64();
-        }
-    }
-
-    public class Buff
-    {
-        protected static Envir Envir
-        {
-            get { return Envir.Main; }
-        }
-
-        private Dictionary<string, object> Data { get; set; } = new Dictionary<string, object>();
-
-        public BuffInfo Info;
-        public MapObject Caster;
-        public uint ObjectID;
-        public long ExpireTime;
-
-        public long LastTime, NextTime;
-
-        public Stats Stats;
-
-        public int[] Values;
-
-        public bool FlagForRemoval;
-        public bool Paused;
-
-        public BuffType Type
-        {
-            get { return Info.Type; }
-        }
-
-        public BuffStackType StackType
-        {
-            get { return Info.StackType; }
-        }
-
-        public BuffProperty Properties
-        {
-            get { return Info.Properties; }
-        }
-
-        public Buff(BuffType type)
-        {
-            Info = Envir.GetBuffInfo(type);
-            Stats = new Stats();
-            Data = new Dictionary<string, object>();
-        }
-
-        public Buff(BinaryReader reader, int version, int customVersion)
-        {
-            var type = (BuffType)reader.ReadByte();
-
-            Info = Envir.GetBuffInfo(type);
-
-            Caster = null;
-
-            if (version < 88)
-            {
-                var visible = reader.ReadBoolean();
-            }
-
-            ObjectID = reader.ReadUInt32();
-            ExpireTime = reader.ReadInt64();
-
-            if (version <= 84)
-            {
-                Values = new int[reader.ReadInt32()];
-
-                for (int i = 0; i < Values.Length; i++)
-                {
-                    Values[i] = reader.ReadInt32();
-                }
-
-                if (version < 88)
-                {
-                    var infinite = reader.ReadBoolean();
-                }
-
-                Stats = new Stats();
-                Data = new Dictionary<string, object>();
-            }
-            else
-            {
-                if (version < 88)
-                {
-                    var stackable = reader.ReadBoolean();
-                }
-
-                Values = new int[0];
-                Stats = new Stats(reader, version, customVersion);
-                Data = new Dictionary<string, object>();
-
-                int count = reader.ReadInt32();
-
-                for (int i = 0; i < count; i++)
-                {
-                    var key = reader.ReadString();
-                    var length = reader.ReadInt32();
-
-                    var array = new byte[length];
-
-                    for (int j = 0; j < array.Length; j++)
-                    {
-                        array[j] = reader.ReadByte();
-                    }
-
-                    Data[key] = Functions.DeserializeFromBytes(array);
-                }
-
-                if (version > 86)
-                {
-                    count = reader.ReadInt32();
-
-                    Values = new int[count];
-
-                    for (int i = 0; i < count; i++)
-                    {
-                        Values[i] = reader.ReadInt32();
-                    }
-                }
-            }
-        }
-
-        public void Save(BinaryWriter writer)
-        {
-            writer.Write((byte)Type);
-            writer.Write(ObjectID);
-            writer.Write(ExpireTime);
-
-            Stats.Save(writer);
-
-            writer.Write(Data.Count);
-
-            foreach (KeyValuePair<string, object> pair in Data)
-            {
-                var bytes = Functions.SerializeToBytes(pair.Value);
-
-                writer.Write(pair.Key);
-                writer.Write(bytes.Length);
-
-                for (int i = 0; i < bytes.Length; i++)
-                {
-                    writer.Write(bytes[i]);
-                }
-            }
-
-            writer.Write(Values.Length);
-
-            for (int i = 0; i < Values.Length; i++)
-            {
-                writer.Write(Values[i]);
-            }
-        }
-
-        public T Get<T>(string key)
-        {
-            if (!Data.TryGetValue(key, out object result))
-            {
-                return default;
-            }
-
-            return (T)result;
-        }
-
-        public void Set(string key, object val)
-        {
-            Data[key] = val;
-        }
-
-        public ClientBuff ToClientBuff()
-        {
-            return new ClientBuff
-            {
-                Type = Type,
-                Caster = Caster?.Name ?? "",
-                ObjectID = ObjectID,
-                Visible = Info.Visible,
-                Infinite = StackType == BuffStackType.Infinite,
-                Paused = Paused,
-                ExpireTime = ExpireTime,
-                Stats = new Stats(Stats),
-                Values = Values
-            };
         }
     }
 }
